Split ci_processing some more.
